requirements.txt
arxiv
langchain-openai
langchain-google-genai
langchain-core
langgraph
python-dotenv
pydantic
fpdf2
